refactoring + testability
