+++ manifest.json
@@ -1,1 +1,1 @@
-{"manifest_version":3,"name":"TubeBouncer","version":"2.1.2","description":"Tries to stop you from watching too many YouTube videos.","icons":{"48":"icons/bouncer-48.jpg"},"browser_specific_settings":{"gecko":{"id":"tube@bouncer.me","strict_min_version":"101.0"}},"content_scripts":[{"matches":["*://www.youtube.com/*"],"js":["src/store.js","src/main.js"]}],"permissions":["storage","activeTab"],"action":{"default_popup":"static/options.html","default_icon":{"48":"icons/bouncer-48.jpg"}}}
+{"manifest_version":3,"name":"TubeBouncer","version":"2.1.2","description":"Tries to stop you from watching too many YouTube videos.","icons":{"48":"icons/bouncer-48.jpg"},"browser_specific_settings":{"gecko":{"id":"tube@bouncer.me","strict_min_version":"101.0"}},"content_scripts":[{"matches":["*://www.youtube.com/*"],"js":["src/store.js","src/ext.js","src/main.js"]}],"permissions":["storage","activeTab"],"action":{"default_popup":"static/options.html","default_icon":{"48":"icons/bouncer-48.jpg"}}}
