+++ package.json
@@ -3,7 +3,8 @@
   "description": "dev release automation",
   "scripts": {
     "lint": "eslint --fix && web-ext lint --firefox-preview",
-    "start": "web-ext run --firefox-preview"
+    "start": "web-ext run --firefox-preview",
+    "test": "ava **/*.test.js"
   },
   "dependencies": {
     "@actions/core": "^1.10.0",
@@ -11,6 +12,7 @@
   },
   "devDependencies": {
     "@eslint/js": "^9.4.0",
+    "ava": "^6.1.3",
     "eslint": "^9.4.0",
     "globals": "^15.4.0"
   }
