+++ src/options_main.js
@@ -1,0 +1,7 @@
+/* global main */
+
+main();
+setInterval(() => {
+  main();
+}, 5000);
+
