+++ static/options.html
@@ -10,6 +10,7 @@
     <!-- scripts load at root static/ -->
     <script src="../src/store.js"></script>
     <script src="../src/options.js"></script>
+    <script src="../src/options_main.js"></script>
   </head>
   <body>
     <!-- dynamic content inserted here -->
